requirements.txt
requests
pandas
numpy
google-generativeai
fastapi
uvicorn
yfinance
